# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Ontology Learning API
  version: 1.3.1
  description: >
    Endpoints compatible with your Hamilton parsers (`parse_concept`, `parse_mapping`).
    Sampling behavior:
      - samples=1 → single result form
      - samples>1 → probabilistic lists with likelihood in [0,1]
servers:
  - url: https://api.example.com
...
paths:
  /databases:
    get:
      tags: [Databases]
      summary: List all databases
      description: Get a list of all registered databases
      responses:
        '200':
          description: List of databases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Database'
    
    post:
      tags: [Databases]
      summary: Create a new database
      description: >
        Create a new database on the configured PostgreSQL server.
        Uses admin credentials from server configuration.
        Optionally execute SQL script after creation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: 
                  type: string
                  description: Display name for the database
                  example: "Sales Database"
                database_name:
                  type: string
                  description: Name of the database in PostgreSQL (auto-generated if not provided)
                  example: "sales_db"
                sql_content:
                  type: string
                  description: Optional SQL script to execute after database creation
                  example: "CREATE TABLE customers (id SERIAL PRIMARY KEY, name VARCHAR(255));"
                create_if_not_exists:
                  type: boolean
                  default: true
                  description: Create the database if it doesn't already exist
      responses:
        '201':
          description: Database created successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Database' }
        '400':
          description: Invalid payload or creation failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /databases/upload:
    post:
      tags: [Databases]
      summary: Upload database SQL script
      description: Legacy endpoint for uploading SQL files. Creates a SQLite database by default.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, sql_file]
              properties:
                name: { type: string }
                sql_file:
                  type: string
                  format: binary
                  description: SQL script (DDL/DML)
      responses:
        '201':
          description: Database registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Database' }
        '400':
          description: Invalid payload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /databases/from-text:
    post:
      tags: [Databases]
      summary: Create database from SQL text
      description: Create a database from raw SQL text content
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [name, sql_content]
              properties:
                name: 
                  type: string
                  description: Database name
                sql_content:
                  type: string
                  description: Raw SQL script
      responses:
        '201':
          description: Database created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Database' }
        '400':
          description: Invalid payload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /databases/{databaseId}:
    get:
      tags: [Databases]
      summary: Get database metadata
      parameters:
        - $ref: '#/components/parameters/DatabaseId'
      responses:
        '200':
          description: Database metadata
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Database' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    
    delete:
      tags: [Databases]
      summary: Delete database metadata
      description: >
        Delete database metadata from the system.
        Note: This does NOT drop the actual database on the PostgreSQL server,
        it only removes the metadata entry.
      parameters:
        - $ref: '#/components/parameters/DatabaseId'
      responses:
        '204':
          description: Database metadata deleted successfully (no content)
        '404':
          description: Database not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /databases/{databaseId}/schema:
    get:
      tags: [Databases]
      summary: Get database schema metadata
      description: >
        Returns detailed schema information including tables, columns, and their data types.
        Useful for exploring the database structure before ontology generation.
      parameters:
        - $ref: '#/components/parameters/DatabaseId'
      responses:
        '200':
          description: Database schema metadata
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DatabaseSchema' }
        '404':
          description: Database not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /mock/databases:
    get:
      tags: [Mock]
      summary: "[MOCK] List all databases"
      description: Mock endpoint that returns a list of fake databases.
      responses:
        '200':
          description: List of mock databases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Database'
    
    post:
      tags: [Mock]
      summary: "[MOCK] Upload database SQL script"
      description: Mock endpoint that returns fake database metadata without processing SQL.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, sql_file]
              properties:
                name: { type: string }
                sql_file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Mock database created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Database' }
...
  /mock/databases/{databaseId}:
    get:
      tags: [Mock]
      summary: "[MOCK] Get database metadata"
      description: Mock endpoint that returns fake database metadata.
      parameters:
        - $ref: '#/components/parameters/DatabaseId'
      responses:
        '200':
          description: Mock database metadata
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Database' }
    
    delete:
      tags: [Mock]
      summary: "[MOCK] Delete database"
      description: Mock endpoint that removes a database from mock storage.
      parameters:
        - $ref: '#/components/parameters/DatabaseId'
      responses:
        '204':
          description: Mock database deleted successfully (no content)
        '404':
          description: Database not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /mock/databases/{databaseId}/schema:
    get:
      tags: [Mock]
      summary: "[MOCK] Get database schema metadata"
      description: Mock endpoint that returns fake schema with sample tables (users, orders, products, etc.).
      parameters:
        - $ref: '#/components/parameters/DatabaseId'
      responses:
        '200':
          description: Mock database schema
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DatabaseSchema' }
...
components:
  parameters:
    DatabaseId:
      name: databaseId
      in: path
      required: true
      schema: { type: string }
      description: Database identifier returned by the upload endpoint.
...
  schemas:
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error: { type: string, example: "NotFoundError" }
        message: { type: string, example: "Database not found." }
    Database:
      type: object
      required: [id, name, createdAt]
      properties:
        id: 
          type: string
          description: Unique database identifier
          example: "db_a1b2c3d4e5"
        name: 
          type: string
          description: Display name for the database
          example: "Sales Database"
        createdAt: 
          type: string
          format: date-time
          description: When the database was registered
        tableCount:
          type: integer
          description: Number of tables in the database
          example: 15
        status:
          type: string
          enum: [pending, connected, failed, disconnected]
          description: Connection status of the database
          example: "connected"
    DatabaseSchema:
      type: object
      required: [databaseId, tables, tableCount]
      properties:
        databaseId: { type: string, description: "Database identifier" }
        tableCount: { type: integer, description: "Total number of tables" }
        tables:
          type: array
          description: "List of tables with their columns"
          items:
            type: object
            required: [schema, name, columns]
            properties:
              schema: { type: string, description: "Schema name (e.g., 'public')" }
              name: { type: string, description: "Table name" }
              columnCount: { type: integer, description: "Number of columns in table" }
              columns:
                type: array
                description: "List of columns in the table"
                items:
                  type: object
                  required: [name, dataType]
                  properties:
                    name: { type: string, description: "Column name" }
                    dataType: { type: string, description: "Column data type" }
                    nullable: { type: boolean, description: "Whether column allows NULL" }
                    isPrimaryKey: { type: boolean, description: "Whether column is part of primary key" }
                    isForeignKey: { type: boolean, description: "Whether column is a foreign key" }
                    defaultValue: { type: string, description: "Default value if any" }
...